Download TalkBank Corpus Data

Downloads the utterances, transcripts, participants, tokens, and token types data from the TalkBank database and saves it to disk in the specified target directory.

Usage

get_talkbank_data(
  corpus_name,
  corpus_path,
  target_dir,
  force = FALSE,
  confirmed = FALSE,
  timeout = 60
)

Arguments

corpus_name

The name of the TalkBank corpus to download data from.

corpus_path

The path to the TalkBank corpus to download data from.

target_dir

The directory to save the downloaded data to.

force

If TRUE, the data will be downloaded even if it already exists on disk.

confirmed

If TRUE, the user has confirmed that they have permission to use the data. If FALSE, the function will prompt the user to confirm permission. Setting this to TRUE is useful for reproducible workflows.

timeout

The maximum time in seconds to wait for the data to be downloaded. Default is 60 seconds.

Value

A message indicating whether the data was acquired or already existed on disk, writes the data files to disk in the specified target directory.

Examples

if (FALSE) { # \dontrun{
data_dir <- file.path(tempdir(), "data")

get_talkbank_data(
 corpus_name = "ca",
 corpus_path = c("ca", "Nahuatl"),
 target_dir = data_dir,
 confirmed = TRUE)
} # }